Create console/board_display which is concerned with the appearance of the board List
@maikon
@jsuchy
@felipesere
